requirements.txt
pandas
Pillow
openpyxl
matplotlib
vaderSentiment
networkx
reverse_geocoder
timezonefinder
pytest